user-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
sqlmodel>=0.0.16
sqlalchemy>=2.0
requests
